Combine the about and experience sections
Combined the about and experience sections to form a profile section,
with a bento box ios style, and the icons running down the right side
of the page. Also removed the alternating format of the portfolio
section as it was formatting badly on mobiles. Currently looks good one
sided with the images slightly larger. Still need to add responsiveness
to the profile section but am satisfied with the desktop version.